+++ lib/esbonio/esbonio/server/cli.py
@@ -71,7 +71,9 @@ def main(argv: Optional[Sequence[str]] = None):
         "esbonio.server.features.directives",
         "esbonio.server.features.roles",
         "esbonio.server.features.rst.directives",
+        "esbonio.server.features.rst.roles",
         "esbonio.server.features.myst.directives",
+        "esbonio.server.features.myst.roles",
         "esbonio.server.features.sphinx_support.diagnostics",
         "esbonio.server.features.sphinx_support.symbols",
         "esbonio.server.features.sphinx_support.directives",
